Updated rest package structure for api interface for Swagger
